Team changes

+++ controller/TeamController.js
@@ -2,7 +2,6 @@ const passport = require("passport");
 const Club = require("../models/ClubModel");
 const Team = require("../models/TeamModel");
 const User = require("../models/UserModel");
-//const Team = require("../models/MasterModel");
 
 class TeamController{
     
@@ -10,26 +9,19 @@ class TeamController{
 
         if(req.isAuthenticated()){
 
-            const club = await Club.findOne({
-                raw: true,
-                where: {
-                    fkUserId: req.user.id
-                }
-            });
-            
             Team.findAll({
                 raw: true,
-                where: {
-                    fkClubId: club.id,
+                include: {
+                    model: Club,
+                    where: {
+                        UserId: req.user.id,
+                    }
                 }
             }).then(function(teams){
-                //console.log(teams)
                 res.render("equipes", { teams: teams});
             }).catch(function(err){
                 res.send(err);
             });
-            
-            
         }else{
             res.redirect("/login");
         }
@@ -39,42 +31,41 @@ class TeamController{
 
         if(req.isAuthenticated()){
 
-            const {teamName, teamType, trainingWeekDay, trainingStartAt, courtName, address, isActive} = req.body;
-            
-            const clubId = await Club.findOne({
+            let {teamName, teamType, trainingWeekDay, trainingStartAt, courtName, address, isActive} = req.body;
+
+            if(isActive == 'on'){
+                isActive = true;
+            }else{
+                isActive = false;
+            }
+
+           Club.findOne({
                 raw: true,
                 where: {
-                    fkUserId: req.user.id,
+                    UserId: req.user.id
                 }
-            })
-
-            const team = {
-                teamName,
-                teamType,
-                trainingWeekDay,
-                trainingStartAt,
-                courtName,
-                address,
-                isActive,
-                fkClubId: clubId.id,
-            }
-
-            if(team.isActive == 'on'){
-                team.isActive = true;
-            }else{
-                team.isActive = false;
-            }
-
-            await Team.create(team).then(function(team){
-                if(team){
-                    res.status(200).redirect("/adm/equipes");
-                }else{
-                    res.send("Not saved");
-                }
+            }).then(function(club){
+                Team.create({
+                    teamName,
+                    teamType,
+                    trainingWeekDay,
+                    trainingStartAt,
+                    courtName,
+                    address,
+                    isActive,
+                    ClubId: club.id,
+                }).then(function(team){
+                    if(team){
+                        console.log(team);
+                        res.status(200).redirect("/adm/equipes");
+                    }else{
+                        res.send("Not saved");
+                    }
+                }).catch(function(err){
+                    console.log(err);
+                })
             }).catch(function(err){
-                if(err){
-                    res.send(err);
-                }
+                res.send(err);
             });
         }
     }
@@ -84,28 +75,46 @@ class TeamController{
         const {deleteTeam} = req.body;
 
         if(req.isAuthenticated()){
-            const isDeleted = await Team.destroy({
+
+            Team.findAll({
+                raw: true,
                 where: {
                     id: deleteTeam
                 }
+            }).then(function(team){
+                console.log(team[0].ClubId);
+                Club.findOne({
+                    raw: true,
+                    where: {
+                        id: team[0].ClubId,
+                    }
+                }).then(function(club){
+                    console.log(club);
+                    if(req.user.id == club.UserId){
+                        Team.destroy({
+                            where: {
+                                id: deleteTeam,
+                            }
+                        }).then(function(){ res.redirect("/adm/equipes")} ).catch(function(err){res.send(err)})
+                    }else{
+                        console.log("Cannot Delete");
+                    }
+                }).catch((err) => {
+                    res.send(err)
+                });
+            }).catch(function (err) { 
+                res.send(err);
             });
-            
-            if(isDeleted == 1){
-                res.status(200).redirect("/adm/equipes")
-            }else{
-                res.send("Algo de errado aconteceu");
-            }
         }else{
             res.redirect("/login");
         }
     }
 
-    static async editTeam(req, res){
+    static async updateTeam(req, res){
         
         const { teamid, teamNameEdit, teamTypeEdit, trainingWeedDayEdit, trainingStartAtEdit, courtNameEdit, addressEdit, isActiveEdit } = req.body;
 
         let isActive;
-
 
         if(req.isAuthenticated()){
 
@@ -141,6 +150,12 @@ class TeamController{
             res.redirect("/login")
         }
     }
+
+    /* This method should be implemented to get the xhr 
+    requested data on edit team functionality */
+    // static getTeamById(req, res){
+        
+    // }
 }
 
 module.exports = TeamController;
